fix: types and rest initialization

+++ protocol-units/bridge/config/src/common/movement.rs
@@ -36,7 +36,7 @@ pub struct MovementConfig {
 	#[serde(default = "default_rest_connection_hostname")]
 	pub rest_hostname: String,
 	#[serde(default = "default_rest_connection_port")]
-	pub rest_port: u16,
+	pub rest_port: u32,
 }
 
 // The default private key
@@ -54,7 +54,7 @@ env_default!(
 	DEFAULT_REST_CONNECTION_HOSTNAME.to_string()
 );
 
-env_default!(default_rest_connection_port, "REST_CONNECTION_PORT", u16, 308833);
+env_default!(default_rest_connection_port, "REST_CONNECTION_PORT", u32, 308833);
 
 env_default!(
 	default_movement_native_address,
